util: Handle
Introduce a Handle class as a utility to the project. That is going to
be really useful in the future when we will store elements  of the scene
in vector/pack containers.

A Pack is a vector of elements which works through handles. It has a
list of indices to elements, and a list of free positions into the list
of indices. This creates a level of indirection so that handles point to
indices instead of elements. With this approach we are able to remove
elements from the pack easily while keeping the contiguous property of
the pack.