vkr: Image ready semaphores
This is an important fix which fixes a logic error related to how we
manage image ready semaphores. These can not be associated to specific
frames as we do not know, at image acquisition time, which image index
is going to be returned. The best approach is to just use a ring
buffer, creating semaphores on the fly when needed.