vkr: Fix frame handling
The way we handled frames so far was flawed. We can not use image
indices if at the same time we change the size the vector containing the
frames. To fix this, we do not ever resize the frame vector. Instead, we
take frames replacing them with an empty value until returned.