core: Pipeline handles
Improve pipeline pools by expecting a handle to a pipeline instead of a
weakly typed index. We this we also improve the material structure by
expecting a type which can be converted into a pipeline handle at
construction time. This is a requirement as we need to know how we will
draw that specific material.

Improve pipeline pool implementation by using pipeline handles instead
of simple usize indices. We also implement From and Into traits between
pipeline handles and shader enums.